pyproject.toml
[tool.poetry]
name = "incarn"
version = "0.0.1"
description = "A bot for friends and enemies."
authors = ["SidVeld <sidveld@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
rich = "^13.6.0"
py-cord = "^2.4.1"
pyyaml = "^6.0.1"
tortoise-orm = "^0.20.0"
dacite = "^1.8.1"
asyncpg = "^0.28.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
flake8-quotes = "^3.3.2"
flake8-return = "^1.2.0"
taskipy = "^1.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
